Fix annotating of local variables in nested functions
* Add support for parsing modes, add special mode for parsing left hand
sides of assignments.

* Add `parseScopingVars` a separate parser for first argument of `With`,
`Module` and `Block` taking into account fact that variables on right
hand side of assignments are not local variables.

* Split "PatternName" local variable type to "Rules" and "Assignment"
since they behave differently.

* Change `extractLocalVariableNames` so that it no longer treats
variables found in inner scoping construct as local variables for outer
construct.

* Parse function names without local variables of this function
assigned. So that if symbol being head of expression is also local
variable inside this expression, then head is not annotated as local
variable.

* Syntax types gathering and normalization overhaul. Append types
instead of prepending them.

* Add tests for above changes.